extract more functions into oladapter
